src/requirements.txt
matplotlib
numpy
opencv-python
pillow
scipy
